resolve typecast error in datapath and remove extra file
