+++ build.sbt
@@ -1,29 +1,64 @@
-  // See README.md for license details.
+// See README.md for license details.
 
-ThisBuild / scalaVersion     := "2.12.13"
-ThisBuild / version          := "0.1.0"
-ThisBuild / organization     := "com.github.merledu"
+def scalacOptionsVersion(scalaVersion: String): Seq[String] = {
+  Seq() ++ {
+    // If we're building with Scala > 2.11, enable the compile option
+    //  switch to support our anonymous Bundle definitions:
+    //  https://github.com/scala/bug/issues/10047
+    CrossVersion.partialVersion(scalaVersion) match {
+      case Some((2, scalaMajor: Long)) if scalaMajor < 12 => Seq()
+      case _ => Seq("-Xsource:2.11")
+    }
+  }
+}
 
-lazy val root = (project in file("."))
-  .settings(
-    name := "Scala-Chisel-Learning-Journey",
-    libraryDependencies += ("edu.berkeley.cs" %% "chiseltest" % "0.3.2"),
-    libraryDependencies ++= Seq(
-      "edu.berkeley.cs" %% "chisel3" % "3.4.3",
-      "edu.berkeley.cs" %% "chiseltest" % "0.3.3" % "test"
+def javacOptionsVersion(scalaVersion: String): Seq[String] = {
+  Seq() ++ {
+    // Scala 2.12 requires Java 8. We continue to generate
+    //  Java 7 compatible code for Scala 2.11
+    //  for compatibility with old clients.
+    CrossVersion.partialVersion(scalaVersion) match {
+      case Some((2, scalaMajor: Long)) if scalaMajor < 12 =>
+        Seq("-source", "1.7", "-target", "1.7")
+      case _ =>
+        Seq("-source", "1.8", "-target", "1.8")
+    }
+  }
+}
 
-    ),
+name := "Auqaab-Single-Cycle-Core"
 
-    scalacOptions ++= Seq(
-      "-Xsource:2.11",
-      "-language:reflectiveCalls",
-      "-deprecation",
-      "-feature",
-      "-Xcheckinit",
-      // Enables autoclonetype2 in 3.4.x (on by default in 3.5)
-      "-P:chiselplugin:useBundlePlugin"
-    ),
-    addCompilerPlugin("edu.berkeley.cs" % "chisel3-plugin" % "3.4.3" cross CrossVersion.full),
-    addCompilerPlugin("org.scalamacros" % "paradise" % "2.1.1" cross CrossVersion.full)
+version := "3.2.0"
+
+scalaVersion := "2.12.10"
+
+crossScalaVersions := Seq("2.12.10", "2.11.12")
+
+
+resolvers ++= Seq(
+  Resolver.sonatypeRepo("snapshots"),
+  Resolver.sonatypeRepo("releases")
+)
+
+// Provide a managed dependency on X if -DXVersion="" is supplied on the command line.
+val defaultVersions = Map(
+  "chisel3" -> "3.4.2",
+  "chisel-iotesters" -> "1.5.0"
   )
 
+libraryDependencies ++= Seq("chisel3","chisel-iotesters").map {
+  dep: String => "edu.berkeley.cs" %% dep % sys.props.getOrElse(dep + "Version", defaultVersions(dep)) }
+
+libraryDependencies += "edu.berkeley.cs" %% "chiseltest" % "0.3.2" % "test"
+
+libraryDependencies += "org.scalatest" %% "scalatest" % "3.0.5" % "test"
+
+logBuffered in Test := false
+
+parallelExecution in Test := false
+
+scalacOptions ++= scalacOptionsVersion(scalaVersion.value)
+
+javacOptions ++= javacOptionsVersion(scalaVersion.value)
+
+trapExit := false
